[Extra Credit] Use camelcase for user properties
If we want to use camelCase for a more Kotlin-native feel for our
user, we update the property name everywhere, run the tests, and
see them fail because the property ends up being `null`. However,
adding the `@Json(name="")` annotation, we will get the desired
results, and Moshi will map the properties from the specified JSON
key.